Migrate new stories users handler to BackendSaga
Done to align with where the action creator is located.
